Move base_suite to azureiam
